Move TopicDeferredPublish SDK client from draft to topic.
