+++ ydb/public/sdk/cpp/src/client/topic/impl/ydb_topic_deferred_publish.cpp
@@ -1,4 +1,4 @@
-#include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/draft/ydb_topic_deferred_publish.h>
+#include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/topic/ydb_topic_deferred_publish.h>
 
 #include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/types/operation/operation.h>
