+++ ydb/public/sdk/cpp/src/client/topic/ut/topic_deferred_publish_ut.cpp
@@ -1,6 +1,6 @@
 #include "ut_utils/topic_sdk_test_setup.h"
 
-#include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/draft/ydb_topic_deferred_publish.h>
+#include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/topic/ydb_topic_deferred_publish.h>
 #include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/driver/driver.h>
 #include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/topic/client.h>
 #include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/types/tx/tx.h>
